Changes the mandatory parameters to update subscription route
It's not mandatory to send all parameters to the update. It's possible to change,
for instance, only the plan_id without to alter the card or the payment method.
To implement it, we use the "memento pattern" instead to make a new request of the early data subscription.

Changes the "plan_id" parameter name. Use "plan_id" instead of "plan" because the body
of request with "plan" is not accepted by API in the route PUT /subscriptions/:id.

Create methods to load the card, plan and payment method with specific validations
to be attached in the body of the request.

Changes unit tests and creates behaviour tests.